Create _Volume base class
